Eliminate dynamic creation of serial test cases
